requirement.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
numpy==1.24.2
Pillow==9.4.0
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
